requirements.txt
Flask
flask-pymongo
pymongo
flask-cors
python-dotenv
dnspython
gunicorn